fix posts with my vote relationship to bind vote color

+++ app/Http/Controllers/Api/V1/VoteController.php
@@ -14,6 +14,7 @@ class VoteController extends Controller
 {
     public function upVote(Post $post)
     {
+        $status = Post::UNVOTE;
         $vote = Vote::where([
             'post_id' => $post->id,
             'user_id' => Auth::id()
@@ -22,15 +23,12 @@ class VoteController extends Controller
         /** if vote already exist */
         if ($vote) {
             /** check if vote is upvote (1) then delete */
-            if ($vote->vote === 1) {
-                $vote->delete();
-                $post->update(['total_votes' => ($post->total_votes - 1)]);
-            }
+            if ($vote->vote === 1) $vote->delete();
 
             /** check if vote is downvote (-1) then update to upvote (1) */
             if ($vote->vote === -1) {
                 $vote->update(['vote' => 1]);
-                $post->update(['total_votes' => ($post->total_votes + 1)]);
+                $status = Post::UPVOTE;
             }
         } else {
             /** create new upvote entry */
@@ -40,14 +38,21 @@ class VoteController extends Controller
                 'vote' => 1
             ]);
 
-            $post->update(['total_votes' => ($post->total_votes + 1)]);
+            $status = Post::UPVOTE;
         }
 
-        return $this->customResponse('success');
+        /** update post's total votes */
+        $post->update(['total_votes' => $post->votes->sum('vote')]);
+
+        return $this->customResponse('success', [
+            'total_votes' => $post->total_votes,
+            'status' => $status
+        ]);
     }
 
     public function downVote(Post $post)
     {
+        $status = Post::UNVOTE;
         $vote = Vote::where([
             'post_id' => $post->id,
             'user_id' => Auth::id()
@@ -56,15 +61,12 @@ class VoteController extends Controller
         /** if vote already exist */
         if ($vote) {
             /** check if vote is downvote (-1) then delete */
-            if ($vote->vote === -1) {
-                $vote->delete();
-                $post->update(['total_votes' => ($post->total_votes + 1)]);
-            }
+            if ($vote->vote === -1) $vote->delete();
 
             /** check if vote is upvote (1) then update to downvote (-1) */
             if ($vote->vote === 1) {
                 $vote->update(['vote' => -1]);
-                $post->update(['total_votes' => ($post->total_votes - 1)]);
+                $status = Post::DOWNVOTE;
             }
         } else {
             /** create new upvote entry */
@@ -74,9 +76,15 @@ class VoteController extends Controller
                 'vote' => -1
             ]);
 
-            $post->update(['total_votes' => ($post->total_votes - 1)]);
+            $status = Post::DOWNVOTE;
         }
 
-        return $this->customResponse('success');
+        /** update post's total votes */
+        $post->update(['total_votes' => $post->votes->sum('vote')]);
+
+        return $this->customResponse('success', [
+            'total_votes' => $post->total_votes,
+            'status' => $status
+        ]);
     }
 }
